+++ rust-engine/src/ast/identifiers/global_id.rs
@@ -175,6 +175,14 @@ impl GlobalId {
             _ => None,
         }
     }
+
+    /// Returns `true` if self is a tuple constructor or a tuple type.
+    // TODO: this is a hack, see issue https://github.com/cryspen/hax/issues/1671.
+    pub fn is_tuple(&self) -> bool {
+        self.to_debug_string()
+            .strip_prefix("hax_Tuple")
+            .is_some_and(|value| value.parse::<usize>().is_ok())
+    }
 }
 
 impl ConcreteId {