+++ rust-engine/src/resugarings.rs
@@ -11,6 +11,41 @@ use crate::ast::*;
 use crate::printer::*;
 use std::collections::HashSet;
 
+/// Transforms [`ItemKind::Fn`] of arity zero into [`ResugaredItemKind::Constant`].
+/// Rust `const` items are encoded by the `ImportThir` phase of the hax engine as function of arity zero.
+/// Functions of arity zero themselves are encoded as functions operating on one argument of type `()`.
+#[derive(Copy, Clone, Default)]
+pub struct FunctionsToConstants;
+
+impl AstVisitorMut for FunctionsToConstants {
+    fn enter_item_kind(&mut self, item_kind: &mut ItemKind) {
+        let ItemKind::Fn {
+            name,
+            generics,
+            body,
+            params,
+            safety: SafetyKind::Safe,
+        } = item_kind
+        else {
+            return;
+        };
+        if !params.is_empty() {
+            return;
+        }
+        *item_kind = ItemKind::Resugared(ResugaredItemKind::Constant {
+            name: name.clone(),
+            body: body.clone(),
+            generics: generics.clone(),
+        });
+    }
+}
+
+impl Resugaring for FunctionsToConstants {
+    fn name(&self) -> String {
+        "functions-to-constants".to_string()
+    }
+}
+
 /// Binop resugaring. Used to identify expressions of the form `(f e1 e2)` where
 /// `f` is a known identifier.
 pub struct BinOp {
@@ -63,5 +98,50 @@ impl AstVisitorMut for BinOp {
 impl Resugaring for BinOp {
     fn name(&self) -> String {
         "binop".to_string()
+    }
+}
+
+/// Tuples resugaring. Resugars tuple constructors to the dedicated expression variant [`ResugaredExprKind::Tuple`],
+/// and tuple types to the dedicated type variant [`ResugaredTyKind::Tuple`].
+pub struct Tuples;
+
+impl AstVisitorMut for Tuples {
+    fn enter_expr_kind(&mut self, x: &mut ExprKind) {
+        let (constructor, fields) = match x {
+            ExprKind::Construct {
+                constructor,
+                is_record: false,
+                is_struct: true,
+                base: None,
+                fields,
+            } => (constructor, &fields[..]),
+            ExprKind::GlobalId(constructor) => (constructor, &[][..]),
+            _ => return,
+        };
+        if constructor.is_tuple() {
+            let args = fields.iter().map(|(_, e)| e).cloned().collect();
+            *x = ExprKind::Resugared(ResugaredExprKind::Tuple(args))
+        }
+    }
+    fn enter_ty_kind(&mut self, x: &mut TyKind) {
+        let TyKind::App { head, args } = x else {
+            return;
+        };
+        if head.is_tuple() {
+            let Some(args) = args
+                .iter()
+                .map(GenericValue::expect_ty)
+                .collect::<Option<Vec<_>>>()
+            else {
+                return;
+            };
+            *x = TyKind::Resugared(ResugaredTyKind::Tuple(args.into_iter().cloned().collect()))
+        }
+    }
+}
+
+impl Resugaring for Tuples {
+    fn name(&self) -> String {
+        "tuples".to_string()
     }
 }