+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Changes to the Rust Engine:
  - The module `names` now produces `ExplicitDefId`s instead of `DefId`s (#1648)
+ - Add a resugaring `FunctionsToConstants` (#1559)
+ - Drop the tuple nodes of the AST, add resugaring node for tuples (#1662)
+
+Miscellaneous:
+ - A lean tutorial has been added to the hax website.
 
 ## 0.3.4